+++ src/main/java/NetworkAPIPrototype.java
@@ -1,9 +1,9 @@
-package annotations;
+package main.java;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Target;
 
 @Target(ElementType.METHOD)
-public @interface ProcessAPIPrototype {
+public @interface NetworkAPIPrototype {
 	// Marker annotation, should be applied to a method within a prototype class
 }
