--- test/apiSmokeTest1.java
@@ -1,4 +1,0 @@
-
-public class apiSmokeTest1 {
-
-}
